requirements.txt
appengine-python-standard
asgiref
astroid
autopep8
black
click
dill
Django
djangorestframework
python-dotenv
imagekitio
isort
lazy-object-proxy
mccabe
mypy-extensions
pathspec
platformdirs
pytz
sqlparse
tomli
tomlkit
typing_extensions
wrapt

